Split parameter offload from z3
